safety: split shared data to honor file size cap
